# internal/keeper/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Keeper API",
    "description": "API spec for Tacito Square Keeper component, managing agent configurations, communities, and support entities.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Primary API Version 1"
    }
  ],
...
  "paths": {
...
    "/prompts": {
      "post": {
        "summary": "Create a new Prompt Template version 1",
        "operationId": "createPromptTemplate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePromptTemplateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created Prompt Template",
            "headers": {
              "Location": {
                "description": "The URI of the newly created prompts",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request / Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "agent-config/prompt-engineering"
        ]
      },
      "get": {
        "summary": "List unique Prompt Templates (highest version of each name)",
        "operationId": "listPromptTemplates",
        "responses": {
          "200": {
            "description": "List of Prompt Templates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PromptTemplate"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "agent-config/prompt-engineering"
        ]
      }
    },
...
    "/prompt-collections": {
      "post": {
        "summary": "Create a new Prompt Collection",
        "operationId": "createPromptCollection",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePromptCollectionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created Prompt Collection",
            "headers": {
              "Location": {
                "description": "The URI of the newly created prompt collections",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request / Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "agent-config/prompt-engineering"
        ]
      },
      "get": {
        "summary": "List all configured Prompt Collections",
        "operationId": "listPromptCollections",
        "responses": {
          "200": {
            "description": "List of Prompt Collections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PromptCollection"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "agent-config/prompt-engineering"
        ]
      }
    },
...
    "/prompt-collections/{id}/resolve": {
      "get": {
        "summary": "Resolve collection templates (highest active version of each name)",
        "operationId": "resolveCollectionPrompts",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The UUID of the Prompt Collection",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resolved Prompt Templates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PromptTemplate"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "agent-config/prompt-engineering"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePromptTemplateRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "system-behavior"
          },
          "content": {
            "type": "string",
            "example": "You are a helpful AI assistant."
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "active",
              "archived"
            ],
            "example": "active"
          }
        }
      },
...
      "PromptTemplate": {
        "type": "object",
        "required": [
          "id",
          "name",
          "content",
          "status",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "active",
              "archived"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreatePromptCollectionRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "agent-a-prompts"
          },
          "description": {
            "type": "string",
            "example": "A collection of greeting and behavioral templates."
          },
          "templates": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "example": [
              "550e8400-e29b-41d4-a716-446655440000"
            ]
          }
        }
      },
...
      "PromptCollection": {
        "type": "object",
        "required": [
          "id",
          "name",
          "templates",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "templates": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "name is required"
          }
        }
      },
...
    }
  }
}